+++ component/discovery/puppetdb/puppetdb.go
@@ -8,15 +8,17 @@ import (
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/common/config"
 	"github.com/grafana/agent/component/discovery"
+	"github.com/grafana/agent/internal/featuregate"
 	"github.com/prometheus/common/model"
 	prom_discovery "github.com/prometheus/prometheus/discovery/puppetdb"
 )
 
 func init() {
 	component.Register(component.Registration{
-		Name:    "discovery.puppetdb",
-		Args:    Arguments{},
-		Exports: discovery.Exports{},
+		Name:      "discovery.puppetdb",
+		Stability: featuregate.StabilityStable,
+		Args:      Arguments{},
+		Exports:   discovery.Exports{},
 
 		Build: func(opts component.Options, args component.Arguments) (component.Component, error) {
 			return New(opts, args.(Arguments))