+++ component/prometheus/exporter/windows/windows.go
@@ -3,14 +3,16 @@ package windows
 import (
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/prometheus/exporter"
+	"github.com/grafana/agent/internal/featuregate"
 	"github.com/grafana/agent/pkg/integrations"
 )
 
 func init() {
 	component.Register(component.Registration{
-		Name:    "prometheus.exporter.windows",
-		Args:    Arguments{},
-		Exports: exporter.Exports{},
+		Name:      "prometheus.exporter.windows",
+		Stability: featuregate.StabilityStable,
+		Args:      Arguments{},
+		Exports:   exporter.Exports{},
 
 		Build: exporter.New(createExporter, "windows"),
 	})