+++ pkg/flow/declare_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/grafana/agent/internal/featuregate"
 	"github.com/grafana/agent/pkg/flow"
 	"github.com/grafana/agent/pkg/flow/internal/testcomponents"
 	"github.com/grafana/agent/pkg/flow/logging"
@@ -368,7 +369,7 @@ func TestDeclareError(t *testing.T) {
 			}
 			a "example" {}
 			`,
-			expectedError: regexp.MustCompile(`cannot retrieve the definition of component name "b_1"`),
+			expectedError: regexp.MustCompile(`cannot find the definition of component name "b_1"`),
 		},
 		{
 			name: "ForbiddenDeclareLabel",
@@ -384,10 +385,11 @@ func TestDeclareError(t *testing.T) {
 			s, err := logging.New(os.Stderr, logging.DefaultOptions)
 			require.NoError(t, err)
 			ctrl := flow.New(flow.Options{
-				Logger:   s,
-				DataPath: t.TempDir(),
-				Reg:      nil,
-				Services: []service.Service{},
+				Logger:       s,
+				DataPath:     t.TempDir(),
+				MinStability: featuregate.StabilityBeta,
+				Reg:          nil,
+				Services:     []service.Service{},
 			})
 			f, err := flow.ParseSource(t.Name(), []byte(tc.config))
 			require.NoError(t, err)