+++ pkg/flow/internal/controller/component_registry.go
@@ -1,27 +1,70 @@
 package controller
 
-import "github.com/grafana/agent/component"
+import (
+	"fmt"
+
+	"github.com/grafana/agent/component"
+	"github.com/grafana/agent/internal/featuregate"
+)
 
 // ComponentRegistry is a collection of registered components.
 type ComponentRegistry interface {
-	// Get looks up a component by name.
-	Get(name string) (component.Registration, bool)
+	// Get looks up a component by name. It returns an error if the component does not exist or its usage is restricted,
+	// for example, because of the component's stability level.
+	Get(name string) (component.Registration, error)
 }
 
-// DefaultComponentRegistry is the default [ComponentRegistry] which gets
+type defaultComponentRegistry struct {
+	minStability featuregate.Stability
+}
+
+// NewDefaultComponentRegistry creates a new [ComponentRegistry] which gets
 // components registered to github.com/grafana/agent/component.
-type DefaultComponentRegistry struct{}
-
-// Get retrieves a component using [component.Get].
-func (reg DefaultComponentRegistry) Get(name string) (component.Registration, bool) {
-	return component.Get(name)
+func NewDefaultComponentRegistry(minStability featuregate.Stability) ComponentRegistry {
+	return defaultComponentRegistry{
+		minStability: minStability,
+	}
 }
 
-// RegistryMap is a map which implements [ComponentRegistry].
-type RegistryMap map[string]component.Registration
+// Get retrieves a component using [component.Get]. It returns an error if the component does not exist,
+// or if the component's stability is below the minimum required stability level.
+func (reg defaultComponentRegistry) Get(name string) (component.Registration, error) {
+	cr, exists := component.Get(name)
+	if !exists {
+		return component.Registration{}, fmt.Errorf("cannot find the definition of component name %q", name)
+	}
+	if err := featuregate.CheckAllowed(cr.Stability, reg.minStability, fmt.Sprintf("component %q", name)); err != nil {
+		return component.Registration{}, err
+	}
+	return cr, nil
+}
+
+type registryMap struct {
+	registrations map[string]component.Registration
+	minStability  featuregate.Stability
+}
+
+// NewRegistryMap creates a new [ComponentRegistry] which uses a map to store components.
+// Currently, it is only used in tests.
+func NewRegistryMap(
+	minStability featuregate.Stability,
+	registrations map[string]component.Registration,
+) ComponentRegistry {
+
+	return &registryMap{
+		registrations: registrations,
+		minStability:  minStability,
+	}
+}
 
 // Get retrieves a component using [component.Get].
-func (m RegistryMap) Get(name string) (component.Registration, bool) {
-	reg, ok := m[name]
-	return reg, ok
+func (m registryMap) Get(name string) (component.Registration, error) {
+	reg, ok := m.registrations[name]
+	if !ok {
+		return component.Registration{}, fmt.Errorf("cannot find the definition of component name %q", name)
+	}
+	if err := featuregate.CheckAllowed(reg.Stability, m.minStability, fmt.Sprintf("component %q", name)); err != nil {
+		return component.Registration{}, err
+	}
+	return reg, nil
 }